rename select collection page file
