+++ frontend/src/features/collections/index.ts
@@ -6,7 +6,7 @@ import("./collection-create-dialog");
 import("./collection-replay-dialog");
 import("./collection-workflow-list");
 import("./select-collection-access");
-import("./select-collection-start-page");
+import("./select-collection-page");
 import("./share-collection");
 import("./collection-thumbnail");
 import("./edit-dialog/sharing-section");
